Added a test for the history-resetting action
and removed the old, non-working attempt
